author some docs
