+++ book/src/SUMMARY.md
@@ -1,6 +1,12 @@
 # Summary
 
 - [Intro](./intro.md)
-- [Formality](./formality.md)
-    - [Terms](./terms.md)
-    - [Inference rules](./inference_rules.md)
+- [`formality_core`: the Formality system](./formality_core.md)
+    - [Defining your lang](./formality_core/lang.md)
+    - [Defining terms with the `term` macro](./formality_core/terms.md)
+        - [Implementing `Fold` and `Visit` by hand](./formality_core/impl_fold_visit.md)
+        - [Implementing `Parse` by hand](./formality_core/impl_parse.md)
+    - [Variables](./formality_core/variables.md)
+    - [Collections](./formality_core/collections.md)
+    - [Judgment functions and inference rules](./formality_core/judgment_fn.md)
+    - [FAQ and troubleshooting](./formality_core/faq.md)
