+++ book/src/faq.md
@@ -1,0 +1,1 @@
+# FAQ and troubleshooting
