+++ book/src/formality_core.md
@@ -1,0 +1,10 @@
+# `formality_core`: the Formality system
+
+`a-mir-formality` is build on the formality core system,
+defined by the `formality_core` crate.
+Formality core is a mildly opnionated series of macros, derives, and types
+that let you write high-level Rust code
+in a way that resembles standard type theory notation.
+Its primary purpose is to be used by a-mir-formality
+but it can be used for other projects.
+
