+++ book/src/formality_core/collections.md
@@ -1,0 +1,1 @@
+# Collections
