+++ book/src/formality_core/variables.md
@@ -1,0 +1,1 @@
+# Variables
