+++ crates/formality-rust/src/lib.rs
@@ -1,5 +1,7 @@
+// ANCHOR: use_rust_language
 // Defines the language used by derive(term) and friends.
 use formality_types::rust::FormalityLang;
+// ANCHOR_END: use_rust_language
 
 pub mod grammar;
 pub mod prove;
