+++ crates/formality-types/src/lib.rs
@@ -1,9 +1,8 @@
 #![allow(dead_code)]
 
-pub mod fold;
 pub mod grammar;
-pub mod visit;
 
+// ANCHOR: declare_rust_language
 formality_core::declare_language! {
     pub mod rust {
         const NAME = "Rust";
@@ -13,5 +12,8 @@ formality_core::declare_language! {
         const BINDING_CLOSE = '>';
     }
 }
+// ANCHOR_END: declare_rust_language
 
+// ANCHOR: use_rust_language
 use crate::rust::FormalityLang;
+// ANCHOR_END: use_rust_language
